Refactoring: extract EndpointServer & TrackerService to separate module for further re-use
